Show list. Delete book. Add book

+++ src/main/java/pl/coderslab/warsztat5katw01/model/Book.java
@@ -1,7 +1,7 @@
 package pl.coderslab.warsztat5katw01.model;
 
 public class Book {
-    private long id;
+    private int id;
     private String isbn;
     private String title;
     private String author;
@@ -12,7 +12,7 @@ public class Book {
 
     }
 
-    public Book(long id, String isbn, String title, String author, String publisher, String type) {
+    public Book(int id, String isbn, String title, String author, String publisher, String type) {
         this.id = id;
         this.isbn = isbn;
         this.title = title;
@@ -21,11 +21,11 @@ public class Book {
         this.type = type;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
